+++ iOverlay/src/mesh/outline/mod.rs
@@ -1,7 +1,5 @@
 mod builder;
-mod builder_join_expand;
 pub mod offset;
 mod section;
-mod builder_join_shrink;
 mod builder_join;
 mod uniq_iter;
